fix: removed previous state/heads from co core
This fundamentally is not working because it bypasses pinning by keeping all states alive forever.
If we artificially ignore the the previous state in pinning this also creates the same problem with encryption mappings.
As we can only map a block if it exists so skipping states before it would not work anymore.

Refs: #77, ebba066e72cf099ec08fb370e2025bc8567646eb